Add ability to link symbols from runner in libraries
Note that any source linked to the runner needs to be
added to the native_simulator target so that it's symbols
are accessible. Any symbols in native_simulator are accessible
in runner too, but any symbols in zephyr itself are hidden. This
was done by zephyr team so that internal main in the firmare
itself does not interfere with the native simulator main.